+++ src/reduct/api/api_server.h
@@ -8,6 +8,7 @@
 #include <string>
 
 #include "reduct/api/callbacks.h"
+#include "reduct/auth/token_auth.h"
 #include "reduct/core/error.h"
 
 namespace reduct::api {
@@ -16,6 +17,7 @@ namespace reduct::api {
  * API handler with all the request callbacks
  */
 class IApiHandler : public IInfoCallback,
+                    public IListStorageCallback,
                     public ICreateBucketCallback,
                     public IGetBucketCallback,
                     public IRemoveBucketCallback,
@@ -29,6 +31,14 @@ class IApiHandler : public IInfoCallback,
  */
 class IApiServer {
  public:
+  /**
+   * Components of the API server
+   */
+  struct Components {
+    std::unique_ptr<IApiHandler> handler;
+    std::unique_ptr<auth::ITokenAuthentication> auth;
+  };
+
   struct Options {
     std::string host;
     int port;
@@ -41,7 +51,7 @@ class IApiServer {
    * @param options
    * @return pointer to the implementation
    */
-  static std::unique_ptr<IApiServer> Build(std::unique_ptr<IApiHandler> handler, Options options);
+  static std::unique_ptr<IApiServer> Build(Components components, Options options);
 
   /**
    * Runs HTTP server